Add thm finalization to remove magic theorems
proof_log is now 'a proof_log, where 'a is initially thm during logging.
Once the log for one proof is complete, it is finalized from
'thm proof_log' to 'src proof_log', where src describes where the thm
came from in the proof (a premise, a hypothesis, etc.).

This makes proof replays more meaningful, as we are not inserting magic
theorems derived from thin air.  However, conversions are still not
treated correctly, and the same goes for Conjuncts_then2_log.